upload-service/requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
boto3==1.26.145
python-multipart==0.0.6
pymongo==4.3.3
pydantic==1.10.9
python-dotenv==1.0.0 